Add iOS Podfile and update JWT expiry parsing
Added Podfile and Podfile.lock for iOS Flutter project to enable CocoaPods integration. Updated JWTService.cs to parse expiry minutes as double instead of int. Modified iOS xcconfig files to include Pods configuration. Added new user model and Home view in Flutter frontend.

+++ Course_code_app/BE_app/Web/Services/JWTService.cs
@@ -14,14 +14,14 @@ public class JwtService : IJwtService
     private readonly string _secretKey;
     private readonly string _issuer;
     private readonly string _audience;
-    private readonly int _expiryMinutes;
+    private readonly double _expiryMinutes;
 
     public JwtService(IConfiguration configuration)
     {
         _secretKey = configuration["JwtSettings:SecretKey"]!;
         _issuer = configuration["JwtSettings:Issuer"]!;
         _audience = configuration["JwtSettings:Audience"]!;
-        _expiryMinutes = int.Parse(configuration["JwtSettings:ExpiryMinutes"]!);
+        _expiryMinutes = double.Parse(configuration["JwtSettings:ExpiryMinutes"]!);
     }
 
     public string GenerateToken(userLoginDTO dto)
